@@ -8,34 +8,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.4",
-    "@angular/common": "^4.4.4",
-    "@angular/compiler": "^4.4.4",
-    "@angular/core": "^4.4.4",
-    "@angular/forms": "^4.4.4",
-    "@angular/http": "^4.4.4",
-    "@angular/platform-browser": "^4.4.4",
-    "@angular/platform-browser-dynamic": "^4.4.4",
-    "@angular/router": "^4.4.4",
-    "@swimlane/ngx-charts": "^6.0.2",
-    "@swimlane/ngx-datatable": "^10.2.3",
-    "@webcomponents/custom-elements": "^1.0.0",
-    "clarity-angular": "^0.10.8",
-    "clarity-icons": "^0.10.8",
-    "clarity-ui": "^0.10.8",
+    "@angular/animations": "5.1.0-beta.0",
+    "@angular/common": "5.1.0-beta.0",
+    "@angular/compiler": "5.1.0-beta.0",
+    "@angular/core": "5.1.0-beta.0",
+    "@angular/forms": "5.1.0-beta.0",
+    "@angular/http": "5.1.0-beta.0",
+    "@angular/platform-browser": "5.1.0-beta.0",
+    "@angular/platform-browser-dynamic": "5.1.0-beta.0",
+    "@angular/router": "5.1.0-beta.0",
+    "@swimlane/ngx-charts": "^6.1.0",
+    "@swimlane/ngx-datatable": "^11.0.3",
+    "@webcomponents/custom-elements": "^1.0.4",
+    "clarity-angular": "^0.10.12",
+    "clarity-icons": "^0.10.12",
+    "clarity-ui": "^0.10.12",
     "core-js": "^2.5.1",
-    "d3": "^4.10.2",
-    "rxjs": "^5.4.3",
+    "d3": "^4.11.0",
+    "rxjs": "^5.5.2",
     "ts-helpers": "^1.1.2",
-    "zone.js": "0.8.17"
+    "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.1",
-    "@angular/compiler-cli": "^4.4.4",
-    "@types/node": "^8.0.26",
+    "@angular/cli": "^1.6.0-beta.0",
+    "@angular/compiler-cli": "5.1.0-beta.0",
+    "@angular/language-service": "5.1.0-beta.0",
+    "@types/node": "~6.0.60",
     "ts-node": "3.3.0",
-    "tslint": "5.7.0",
-    "typescript": "~2.5.2",
-    "webpack": "^3.5.6"
+    "tslint": "5.8.0",
+    "typescript": "^2.6.1"
   }
 }
