@@ -5,6 +5,14 @@ import { LogoComponent  } from './components/logo/logo.component';
 import { LoginService   } from './services/api/login.service';
 import { UserInfoService} from './services/user-info.service';
 
+import 'rxjs/add/operator/filter';
+import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/mergeMap';
+import 'rxjs/add/operator/switchMap';
+import { Subscription } from 'rxjs/Subscription';
+import { Observable } from 'rxjs/Observable';
+import { Observer } from 'rxjs/Observer';
+
 @Component({
   selector   : 'home-comp',
   templateUrl: './home.component.html',
@@ -39,7 +47,7 @@ export class HomeComponent   {
         private loginService:LoginService,
         private userInfoService:UserInfoService
     ) {
-        // This block is to retrieve the data from the routes
+        // This block is to retrieve the data from the routes (routes are defined in app-routing.module.ts)
         router.events
         .filter(event => event instanceof NavigationEnd)
         .map( _ => this.router.routerState.root)
@@ -47,11 +55,11 @@ export class HomeComponent   {
             while (route.firstChild) route = route.firstChild;;
             return route;
         })
-        .flatMap(route => route.data)
+        .mergeMap( route => route.data)
         .subscribe(data => {
-            console.log("Route data===: ",data[0]);
-            this.selectedHeaderItemIndex = data[0].selectedHeaderItemIndex;
-            this.selectedSubNavItemIndex = data[0].selectedSubNavItemIndex;
+            console.log("Route data===: ", data[0]);
+            this.selectedHeaderItemIndex = data[0]?data[0].selectedHeaderItemIndex:-1;
+            this.selectedSubNavItemIndex = data[0]?data[0].selectedSubNavItemIndex:-1;
         });
         this.userName = this.userInfoService.getUserName();
 
