@@ -1,7 +1,8 @@
 import { Injectable, Inject } from '@angular/core';
-import { Http, Headers, Response, Request, RequestOptions, URLSearchParams,RequestMethod } from '@angular/http';
+import { HttpClient, HttpHeaders, HttpResponse, HttpRequest,  HttpParams } from '@angular/common/http';
 import { Router } from '@angular/router';
 import { Observable, ReplaySubject, Subject } from 'rxjs';
+import 'rxjs/add/operator/catch';
 import { UserInfoService, LoginInfoInStorage} from '../user-info.service';
 import { AppConfig } from '../../app-config';
 
@@ -9,12 +10,9 @@ import { AppConfig } from '../../app-config';
 @Injectable()
 export class ApiRequestService {
 
-    private headers:Headers;
-    private requestOptions:RequestOptions;
-
     constructor(
         private appConfig:AppConfig,
-        private http: Http,
+        private http: HttpClient,
         private router:Router,
         private userInfoService:UserInfoService
     ) {}
@@ -22,39 +20,21 @@ export class ApiRequestService {
     /**
      * This is a Global place to add all the request headers for every REST calls
      */
-    appendAuthHeader():Headers {
-        let headers = new Headers({'Content-Type': 'application/json'});
+    getHeaders():HttpHeaders {
+        let headers = new HttpHeaders();
         let token = this.userInfoService.getStoredToken();
-        if (token !==null) {
-            headers.append("Authorization", token);
+        headers = headers.append('Content-Type', 'application/json');
+        if (token !== null) {
+            headers = headers.append("Authorization", token);
         }
         return headers;
     }
 
-    /**
-     * This is a Global place to define all the Request Headers that must be sent for every ajax call
-     */
-    getRequestOptions(requestMethod, url:string, urlParam?:URLSearchParams, body?:Object):RequestOptions {
-        let options = new RequestOptions({
-            headers: this.appendAuthHeader(),
-            method : requestMethod,
-            url    : this.appConfig.baseApiPath + url   //this.api + url,
-        });
-        if (urlParam){
-            options = options.merge({ params: urlParam});
-        }
-        if (body){
-            options = options.merge({body: JSON.stringify(body)});
-        }
-        return options;
-    }
-
-    get(url:string, urlParams?:URLSearchParams):Observable<any>{
+    get(url:string, urlParams?:HttpParams):Observable<any>{
         let me = this;
-        let requestOptions = this.getRequestOptions(RequestMethod.Get, url, urlParams);
-        return this.http.request(new Request(requestOptions))
-            .map(resp => resp.json())
+        return this.http.get(this.appConfig.baseApiPath + url, {headers:this.getHeaders(),  params:urlParams} )
             .catch(function(error:any){
+                console.log("Some error in catch");
                 if (error.status === 401 || error.status === 403){
                     me.router.navigate(['/logout']);
                 }
@@ -64,9 +44,7 @@ export class ApiRequestService {
 
     post(url:string, body:Object):Observable<any>{
         let me = this;
-        let requestOptions = this.getRequestOptions(RequestMethod.Post, url, undefined, body);
-        return this.http.request(new Request(requestOptions))
-            .map(resp => resp.json())
+        return this.http.post(this.appConfig.baseApiPath + url, JSON.stringify(body), { headers:this.getHeaders()})
             .catch(function(error:any){
                 if (error.status === 401){
                     me.router.navigate(['/logout']);
@@ -77,9 +55,7 @@ export class ApiRequestService {
 
     put(url:string, body:Object):Observable<any>{
         let me = this;
-        let requestOptions = this.getRequestOptions(RequestMethod.Put, url, undefined, body);
-        return this.http.request(new Request(requestOptions))
-            .map(resp => resp.json())
+        return this.http.put(this.appConfig.baseApiPath + url, JSON.stringify(body), { headers:this.getHeaders()})
             .catch(function(error:any){
                 if (error.status === 401){
                     me.router.navigate(['/logout']);
@@ -90,9 +66,7 @@ export class ApiRequestService {
 
     delete(url:string):Observable<any>{
         let me = this;
-        let requestOptions = this.getRequestOptions(RequestMethod.Delete, url);
-        return this.http.request(new Request(requestOptions))
-            .map(resp => resp.json())
+        return this.http.delete(this.appConfig.baseApiPath + url, { headers:this.getHeaders()})
             .catch(function(error:any){
                 if (error.status === 401){
                     me.router.navigate(['/logout']);
