@@ -1,8 +1,8 @@
 import { Injectable, Inject } from '@angular/core';
-import { Http, Headers, Response, Request, RequestOptions, URLSearchParams,RequestMethod } from '@angular/http';
 import { Observable, ReplaySubject, Subject } from 'rxjs';
 import { ApiRequestService } from './api-request.service';
 import { TranslateService } from './translate.service';
+import { HttpParams} from "@angular/common/http";
 
 @Injectable()
 export class OrderService {
@@ -18,9 +18,9 @@ export class OrderService {
     getOrderInfo(page?:number, size?:number): Observable<any> {
         //Create Request URL params
         let me = this;
-        let params: URLSearchParams = new URLSearchParams();
-        params.set('page', typeof page === "number"? page.toString():"0");
-        params.set('size', typeof page === "number"? size.toString():"1000");
+        let params: HttpParams = new HttpParams();
+        params = params.append('page', typeof page === "number"? page.toString():"0");
+        params = params.append('size', typeof page === "number"? size.toString():"1000");
         let orderListSubject = new Subject<any>(); // Will use this subject to emit data that we want
         this.apiRequest.get('api/orders',params)
             .subscribe(jsonResp => {
@@ -43,9 +43,9 @@ export class OrderService {
     getOrderDetails(orderId:number): Observable<any> {
         //Create Request URL params
         let me = this;
-        let params: URLSearchParams = new URLSearchParams();
+        let params: HttpParams = new HttpParams();
         if (orderId){
-            params.set('orderid', orderId.toString());
+            params = params.append('orderid', orderId.toString());
         }
         let orderDetailSubject = new Subject<any>(); // Will use this subject to emit data that we want
         this.apiRequest.get('api/order-details',params)
